Add decimal, [_]crypt modules and window icon handling in SDL
Fixed sys.path in Python 2.7 not pointing to Python27.zip/lib-tk.
Also fixed the TextIOWrapper in Python 2.7:
Previously it would fail on init with the misleading error message "bad
argument type for built-in operation".
The error was caused by the localemodule unable to find the default
locale on Android, causing getpreferredencoding to return None, which
resulted in the TextIOWrapper passing a null encoding to
PyCodec_IncrementalEncoder and _PyCodec_Lookup which raises the error
bad argument error.
To prevent this, we just return 'ascii' as the preferred encoding, when
we can't determine one, as it is done in Python 3.4.